requirements.txt
openai
gradio
pandas
openpyxl
python-dotenv